first version of nodeag
nodeag is a plugin that replaces incapped deagles with a single pistol
similar to how melee weapons function.  It should function correctly as
long as survivors aren't revived with a defib after dying while incapped
with a deagle.  (They will revive fine, they will just have a pistol
instead of a deagle, a small issue since this is meant for meta which
has no defibs)