More plugin splitting!
l4d_boss_percent is a much simpler plugin now.  All it does is prints
the boss percent either when ready-up goes live or the survivors leave
teh saferoom if ready-up isn't running.  It will also add the boss
percent to the ready panel if it is running my new readyup.

witch_and_tankifier takes all of the spawn control out of
l4d_boss_percent.  It is rewritten so that it can spawn a witch on every
map but will enforce a 5 percent point difference between tank and witch
spawn.  It also has a command to add other maps to the static map trie
since having to recompile a plugin to add one is just bad coding!