Remove client prefs from caster assister
For some reason it doesn't seem to be able to read the values properly.
The function always returns true, but the value is always 0.  Players
have to use the commands to set their speeds every time (or hopefully,
have a config that they run to do it for them).

I tried adding the proper include if anyone cares to look into this in
the future.  Still didn't work.